Set default values properly
Set configured default values correctly with a distinct method.

DMXGateway.set_channels() is flagged as async coroutine and cannot be called
from the constructor of ArtnetLight. Also, send the default values directly
and unconditionally.

Without this, HASS would start up without sending the default values to the
hardware, but reporting them up the stack, so the UI is out of sync.